Split off a Main.scala in c2cpg
